rps: create a base interface for representer point calculators + adapt l2 and LJE
