+++ deel/influenciae/rps/rps_l2.py
@@ -13,13 +13,13 @@ from tensorflow.keras.layers import Input, Dense #pylint:  disable=E0611
 from tensorflow.keras.losses import MeanSquaredError, Loss, Reduction #pylint:  disable=E0611
 from tensorflow.keras.regularizers import L2 #pylint:  disable=E0611
 
-from ..common import BaseInfluenceCalculator
+from . import BaseRepresenterPoint
 from ..types import Tuple, Callable, Union
 
-from ..utils import assert_batched_dataset, BacktrackingLineSearch, dataset_size
+from ..utils import BacktrackingLineSearch, dataset_size
 
 
-class RepresenterPointL2(BaseInfluenceCalculator):
+class RepresenterPointL2(BaseRepresenterPoint):
     """
     A class implementing a method to compute the influence of training points through
     the representer point theorem for kernels.
@@ -59,157 +59,16 @@ class RepresenterPointL2(BaseInfluenceCalculator):
             lambda_regularization: float,
             scaling_factor: float = 0.1,
             epochs: int = 100,
-            layer_index: int = -2,
+            layer_index: int = -1,
     ):
-        assert_batched_dataset(train_set)
-        if hasattr(loss_function, 'reduction'):
-            assert loss_function.reduction == Reduction.NONE
-        self.loss_function = loss_function
+        super().__init__(model, train_set, loss_function, layer_index)
         self.n_train = dataset_size(train_set)
-        self.feature_extractor = Model(inputs=model.input, outputs=model.layers[layer_index].output)
-        self.model = model
         self.train_set = train_set
         self.lambda_regularization = lambda_regularization
         self.scaling_factor = scaling_factor
         self.epochs = epochs
         self.linear_layer = None
         self._train_last_layer(self.epochs)
-
-    def _compute_influence_vector(self, train_samples: Tuple[tf.Tensor, ...]) -> tf.Tensor:
-        """
-        Compute an equivalent of the influence vector for a sample of training points.
-
-        Disclaimer: this vector is not an estimation of the difference between the actual
-        model and the perturbed model without the samples (like it is the case with what is
-        calculated using deel.influenciae.influence).
-
-        Parameters
-        ----------
-        train_samples
-            A tensor with a group of training samples of which we wish to compute the influence.
-
-        Returns
-        -------
-        influence_vectors
-            A tensor with a concatenation of the alpha weights and the feature maps for each sample.
-            This allows for optimizations to be put in place but is not really an influence vector
-            of any kind.
-        """
-        x_batch = self.feature_extractor(train_samples[:-1])
-        alpha = self._compute_alpha(x_batch, train_samples[-1])
-
-        return alpha, x_batch
-
-    def _preprocess_samples(self, samples: Tuple[tf.Tensor, ...]) -> tf.Tensor:
-        """
-        Preprocess a single batch of samples.
-
-        Parameters
-        ----------
-        samples
-            A single batch of tensors containing the samples.
-
-        Returns
-        -------
-        evaluate_vect
-            The preprocessed sample
-        """
-        x_batch = self.feature_extractor(samples[:-1])
-        y_t = tf.argmax(self.model(samples[:-1]), axis=1)
-
-        return x_batch, y_t
-
-    def _estimate_individual_influence_values_from_batch(
-            self,
-            train_samples: Tuple[tf.Tensor, ...],
-            samples_to_evaluate: Tuple[tf.Tensor, ...]
-    ) -> tf.Tensor:
-        """
-        Estimate the (individual) influence scores of a single batch of samples with respect to
-        a batch of samples belonging to the model's training dataset.
-
-        Parameters
-        ----------
-        train_samples
-            A single batch of training samples (and their target values).
-        samples_to_evaluate
-            A single batch of samples of which we wish to compute the influence of removing the training
-            samples.
-
-        Returns
-        -------
-        A tensor containing the individual influence scores.
-        """
-        return self._estimate_influence_value_from_influence_vector(
-            self._preprocess_samples(samples_to_evaluate),
-            self._compute_influence_vector(train_samples)
-        )
-
-    def _estimate_influence_value_from_influence_vector(
-            self,
-            preproc_test_sample: tf.Tensor,
-            influence_vector: tf.Tensor
-    ) -> tf.Tensor:
-        """
-        Compute the influence score for a (batch of) preprocessed test sample(s) and a training "influence vector".
-
-        Parameters
-        ----------
-        preproc_test_sample
-            A tensor with a pre-processed sample to evaluate.
-        influence_vector
-            A tensor with the training influence vector.
-
-        Returns
-        -------
-        influence_values
-            A tensor with influence values for the (batch of) test samples.
-        """
-        # Extract the different information inside the tuples
-        feature_maps_test, labels_test = preproc_test_sample
-        alpha, feature_maps_train = influence_vector
-
-        if len(alpha.shape) == 1 or (len(alpha.shape) == 2 and alpha.shape[1] == 1):
-            influence_values = alpha * tf.matmul(feature_maps_train, feature_maps_test, transpose_b=True)
-        else:
-            influence_values = tf.gather(alpha, labels_test, axis=1, batch_dims=1) * \
-                               tf.matmul(feature_maps_train, feature_maps_test, transpose_b=True)
-        influence_values = tf.transpose(influence_values)
-
-        return influence_values
-
-    def _compute_influence_value_from_batch(self, train_samples: Tuple[tf.Tensor, ...]) -> tf.Tensor:
-        """
-        Compute the influence score for a batch of training samples (i.e. self-influence).
-
-        Parameters
-        ----------
-        train_samples
-            A tensor containing a batch of training samples.
-
-        Returns
-        -------
-        influence_values
-            A tensor with the self-influence of the training samples.
-        """
-        x_batch = self.feature_extractor(train_samples[:-1])
-        alpha = self._compute_alpha(x_batch, train_samples[-1])
-
-        # If the problem is binary classification, take all the alpha values
-        # If multiclass, take only those that correspond to the prediction
-        out_shape = self.model.output_shape
-        if len(out_shape) == 1:
-            influence_values = alpha
-        elif len(out_shape) == 2 and out_shape[1] == 1:
-            influence_values = alpha
-        else:
-            if len(out_shape) > 2:
-                indices = tf.argmax(tf.squeeze(self.model(train_samples[:-1]), axis=-1), axis=1)
-            else:
-                indices = tf.argmax(self.model(train_samples[:-1]), axis=1)
-            influence_values = tf.gather(alpha, indices, axis=1, batch_dims=1)
-
-        return tf.abs(influence_values)
 
     def _train_last_layer(self, epochs: int):
         """
